+++ cli/CHANGELOG.md
@@ -14,6 +14,11 @@ launch the browser after testing type selection in the GUI. Addresses
 [#22003](https://github.com/cypress-io/cypress/issues/22003).
 Addressed in [#28538](https://github.com/cypress-io/cypress/pull/28538).
 
+**Bugfixes:**
+
+- Fixed and issue where cypress launch arguments were not being escaped correctly with multiple values inside quotes. Fixes [#27454](https://github.com/cypress-io/cypress/issues/27454).
+
+
 ## 13.7.2
 
 _Released 4/2/2024_